Adding return types, moving sprite stuff to another file
